requirements.txt
# Dependencias principales
pandas>=2.0.0
streamlit>=1.24.0
flask>=2.2.0
langchain>=0.1.0
openai>=0.27.0
requests>=2.28.0
python-dotenv>=1.0.0
jupyter>=1.0.0

# Dev
black==23.9.1
flake8==6.1.0
pytest==7.4.0
pre-commit==3.4.0
ipython
